requirements.txt
angles
onvif-zeep
scapy
paho-mqtt